@@ -116,7 +116,7 @@
     "tslint": "~6.1.0",
     "ttypescript": "^1.5.12",
     "typedoc": "^0.16.4",
-    "typescript": "~3.9.7"
+    "typescript": ">=4.0.0 <4.1.0"
   },
   "typings": "index.d.ts"
 }
