@@ -4,6 +4,7 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
+    "started:emulated": "concurrently -n ng,firebase -c red,yellow \"ng serve -c emulated\" \"firebase emulators:start\"",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
@@ -16,17 +17,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.2.1",
-    "@angular/common": "~10.2.1",
-    "@angular/compiler": "~10.2.1",
-    "@angular/core": "~10.2.1",
+    "@angular/animations": "~11.0.0",
+    "@angular/common": "~11.0.0",
+    "@angular/compiler": "~11.0.0",
+    "@angular/core": "~11.0.0",
     "@angular/fire": "../dist/packages-dist",
-    "@angular/forms": "~10.2.1",
-    "@angular/platform-browser": "~10.2.1",
-    "@angular/platform-browser-dynamic": "~10.2.1",
-    "@angular/platform-server": "~10.2.1",
-    "@angular/router": "~10.2.1",
-    "@angular/service-worker": "^10.2.1",
+    "@angular/forms": "~11.0.0",
+    "@angular/platform-browser": "~11.0.0",
+    "@angular/platform-browser-dynamic": "~11.0.0",
+    "@angular/platform-server": "~11.0.0",
+    "@angular/router": "~11.0.0",
+    "@angular/service-worker": "^11.0.0",
     "@nguniversal/express-engine": "~10.1.0",
     "core-js": "^3.6.5",
     "firebase": "^8.0.0",
@@ -39,34 +40,34 @@
   },
   "devDependencies": {
     "@angular-devkit/architect": "~0.1001.3",
-    "@angular-devkit/build-angular": "~0.1002.0",
-    "@angular/cli": "~10.2.0",
-    "@angular/compiler-cli": "~10.2.1",
-    "@angular/language-service": "~10.2.1",
+    "@angular-devkit/build-angular": "~0.1100.0",
+    "@angular/cli": "~11.0.0",
+    "@angular/compiler-cli": "~11.0.0",
+    "@angular/language-service": "~11.0.0",
     "@firebase/app-types": "^0.6.1",
     "@nguniversal/builders": "^10.1.0",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^14.11.2",
     "codelyzer": "^6.0.0",
+    "concurrently": "^5.3.0",
     "express": "^4.17.1",
     "express-serve-static-core": "^0.1.1",
     "firebase-admin": "^8.13.0",
     "firebase-functions": "^3.11.0",
     "firebase-functions-test": "^0.2.2",
-    "firebase-tools": "^8.11.1",
+    "firebase-tools": "^8.16.1",
     "fuzzy": "^0.1.3",
     "inquirer": "^6.2.2",
     "inquirer-autocomplete-prompt": "^1.0.1",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~5.1.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "ts-node": "~9.0.0",
     "tslint": "~6.1.3",
-    "typescript": "~3.9.7"
+    "typescript": "~4.0.5"
   }
 }
