@@ -6,16 +6,16 @@ import { FirebaseApp } from '@angular/fire';
   template: `
     Hello world!
     {{ firebaseApp.name }}
-    <app-database></app-database>
+    <!-- TODO wrap the Zone issue <app-database></app-database> -->
     <app-firestore></app-firestore>
     <app-storage></app-storage>
     <app-auth></app-auth>
     <app-remote-config></app-remote-config>
-    <!-- TODO get working with Firebase 8 <app-messaging></app-messaging>-->
+    <app-messaging></app-messaging>
+    <app-functions></app-functions>
   `,
   styles: [``]
 })
 export class HomeComponent {
-  constructor(public readonly firebaseApp: FirebaseApp) {
-  }
+  constructor(public readonly firebaseApp: FirebaseApp) {}
 }
