@@ -3,19 +3,19 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  useEmulators: false,
-  production: false,
-  firebase: {
-    apiKey: 'AIzaSyA7CNE9aHbcSEbt9y03QReJ-Xr0nwKg7Yg',
-    authDomain: 'aftest-94085.firebaseapp.com',
-    databaseURL: 'https://aftest-94085.firebaseio.com',
-    projectId: 'aftest-94085',
-    storageBucket: 'aftest-94085.appspot.com',
-    messagingSenderId: '480362569154',
-    appId: '1:480362569154:web:2fe6f75104cdfb82f50a5b',
-    measurementId: 'G-CBRYER9PJR'
-  },
-  vapidKey: 'BIDPctnXHQDIjcOXxDS6qQcz-QTws7bL8v7UPgFnS1Ky5BZL3jS3-XXfxwRHmAUMOk7pXme7ttOBvVoIfX57PEo'
+    useEmulators: true,
+    production: false,
+    firebase: {
+      apiKey: 'AIzaSyA7CNE9aHbcSEbt9y03QReJ-Xr0nwKg7Yg',
+      authDomain: 'aftest-94085.firebaseapp.com',
+      databaseURL: 'https://aftest-94085.firebaseio.com',
+      projectId: 'aftest-94085',
+      storageBucket: 'aftest-94085.appspot.com',
+      messagingSenderId: '480362569154',
+      appId: '1:480362569154:web:2fe6f75104cdfb82f50a5b',
+      measurementId: 'G-CBRYER9PJR'
+    },
+    vapidKey: 'BIDPctnXHQDIjcOXxDS6qQcz-QTws7bL8v7UPgFnS1Ky5BZL3jS3-XXfxwRHmAUMOk7pXme7ttOBvVoIfX57PEo'
 };
 
 /*
