@@ -5,7 +5,8 @@ export type FirebaseOperation = string | firebase.database.Reference | firebase.
 
 export interface AngularFireList<T> {
   query: DatabaseQuery;
-  valueChanges(events?: ChildEvent[]): Observable<T[]>;
+  valueChanges(events?: ChildEvent[], options?: {}): Observable<T[]>;
+  valueChanges<K extends string>(events?: ChildEvent[], options?: {idField: K}): Observable<(T & {[T in K]?: string})[]>;
   snapshotChanges(events?: ChildEvent[]): Observable<SnapshotAction<T>[]>;
   stateChanges(events?: ChildEvent[]): Observable<SnapshotAction<T>>;
   auditTrail(events?: ChildEvent[]): Observable<SnapshotAction<T>[]>;
