@@ -34,6 +34,11 @@ export const ENABLE_PERSISTENCE = new InjectionToken<boolean>('angularfire2.enab
 export const PERSISTENCE_SETTINGS = new InjectionToken<PersistenceSettings | undefined>('angularfire2.firestore.persistenceSettings');
 export const SETTINGS = new InjectionToken<Settings>('angularfire2.firestore.settings');
 
+// SEMVER(7): use Parameters to detirmine the useEmulator arguments
+// type UseEmulatorArguments = Parameters<typeof firebase.firestore.Firestore.prototype.useEmulator>;
+type UseEmulatorArguments = [string, number];
+export const USE_EMULATOR = new InjectionToken<UseEmulatorArguments>('angularfire2.firestore.use-emulator');
+
 /**
  * A utility methods for associating a collection reference with
  * a query.
@@ -46,7 +51,7 @@ export const SETTINGS = new InjectionToken<Settings>('angularfire2.firestore.set
  *  return ref.where('age', '<', 200);
  * });
  */
-export function associateQuery(collectionRef: CollectionReference, queryFn = ref => ref): AssociatedReference {
+export function associateQuery<T>(collectionRef: CollectionReference<T>, queryFn = ref => ref): AssociatedReference<T> {
   const query = queryFn(collectionRef);
   const ref = collectionRef;
   return { query, ref };
@@ -129,7 +134,8 @@ export class AngularFirestore {
     // tslint:disable-next-line:ban-types
     @Inject(PLATFORM_ID) platformId: Object,
     zone: NgZone,
-    @Optional() @Inject(PERSISTENCE_SETTINGS) persistenceSettings: PersistenceSettings | null
+    @Optional() @Inject(PERSISTENCE_SETTINGS) persistenceSettings: PersistenceSettings | null,
+    @Optional() @Inject(USE_EMULATOR) _useEmulator: any,
   ) {
     this.schedulers = new ɵAngularFireSchedulers(zone);
     this.keepUnstableUntilFirst = ɵkeepUnstableUntilFirstFactory(this.schedulers);
@@ -145,6 +151,10 @@ export class AngularFirestore {
       const firestore = app.firestore();
       if (settings) {
         firestore.settings(settings);
+      }
+      const useEmulator: UseEmulatorArguments | null = _useEmulator;
+      if (useEmulator) {
+        firestore.useEmulator(...useEmulator);
       }
       return firestore;
     });
@@ -173,14 +183,14 @@ export class AngularFirestore {
   collection<T>(path: string, queryFn?: QueryFn): AngularFirestoreCollection<T>;
   // tslint:disable-next-line:unified-signatures
   collection<T>(ref: CollectionReference, queryFn?: QueryFn): AngularFirestoreCollection<T>;
-  collection<T>(pathOrRef: string | CollectionReference, queryFn?: QueryFn): AngularFirestoreCollection<T> {
-    let collectionRef: CollectionReference;
+  collection<T>(pathOrRef: string | CollectionReference<T>, queryFn?: QueryFn): AngularFirestoreCollection<T> {
+    let collectionRef: CollectionReference<T>;
     if (typeof pathOrRef === 'string') {
-      collectionRef = this.firestore.collection(pathOrRef);
+      collectionRef = this.firestore.collection(pathOrRef) as firebase.firestore.CollectionReference<T>;
     } else {
       collectionRef = pathOrRef;
     }
-    const { ref, query } = associateQuery(collectionRef, queryFn);
+    const { ref, query } = associateQuery<T>(collectionRef, queryFn);
     const refInZone = this.schedulers.ngZone.run(() => ref);
     return new AngularFirestoreCollection<T>(refInZone, query, this);
   }
@@ -190,9 +200,9 @@ export class AngularFirestore {
    * and an optional query function to narrow the result
    * set.
    */
-  collectionGroup<T>(collectionId: string, queryGroupFn?: QueryGroupFn): AngularFirestoreCollectionGroup<T> {
+  collectionGroup<T>(collectionId: string, queryGroupFn?: QueryGroupFn<T>): AngularFirestoreCollectionGroup<T> {
     const queryFn = queryGroupFn || (ref => ref);
-    const collectionGroup: Query = this.firestore.collectionGroup(collectionId);
+    const collectionGroup: Query<T> = this.firestore.collectionGroup(collectionId) as firebase.firestore.Query<T>;
     return new AngularFirestoreCollectionGroup<T>(queryFn(collectionGroup), this);
   }
 
@@ -205,10 +215,10 @@ export class AngularFirestore {
   doc<T>(path: string): AngularFirestoreDocument<T>;
   // tslint:disable-next-line:unified-signatures
   doc<T>(ref: DocumentReference): AngularFirestoreDocument<T>;
-  doc<T>(pathOrRef: string | DocumentReference): AngularFirestoreDocument<T> {
-    let ref: DocumentReference;
+  doc<T>(pathOrRef: string | DocumentReference<T>): AngularFirestoreDocument<T> {
+    let ref: DocumentReference<T>;
     if (typeof pathOrRef === 'string') {
-      ref = this.firestore.doc(pathOrRef);
+      ref = this.firestore.doc(pathOrRef) as firebase.firestore.DocumentReference<T>;
     } else {
       ref = pathOrRef;
     }
