@@ -2,13 +2,13 @@ import { Subscriber } from 'rxjs';
 import firebase from 'firebase/app';
 
 export type Settings =  firebase.firestore.Settings;
-export type CollectionReference = firebase.firestore.CollectionReference;
-export type DocumentReference = firebase.firestore.DocumentReference;
+export type CollectionReference<T = DocumentData> = firebase.firestore.CollectionReference<T>;
+export type DocumentReference<T = DocumentData> = firebase.firestore.DocumentReference<T>;
 export type PersistenceSettings = firebase.firestore.PersistenceSettings;
 export type DocumentChangeType = firebase.firestore.DocumentChangeType;
 export type SnapshotOptions = firebase.firestore.SnapshotOptions;
 export type FieldPath = firebase.firestore.FieldPath;
-export type Query = firebase.firestore.Query;
+export type Query<T = DocumentData> = firebase.firestore.Query<T>;
 
 export type SetOptions = firebase.firestore.SetOptions;
 export type DocumentData = firebase.firestore.DocumentData;
@@ -49,14 +49,14 @@ export interface Action<T> {
 }
 
 export interface Reference<T> {
-  onSnapshot: (sub: Subscriber<any>) => any;
+  onSnapshot: (options: firebase.firestore.SnapshotListenOptions, sub: Subscriber<any>) => any;
 }
 
 // A convience type for making a query.
 // Example: const query = (ref) => ref.where('name', == 'david');
-export type QueryFn = (ref: CollectionReference) => Query;
+export type QueryFn<T = DocumentData> = (ref: CollectionReference<T>) => Query<T>;
 
-export type QueryGroupFn = (query: Query) => Query;
+export type QueryGroupFn<T = DocumentData> = (query: Query<T>) => Query<T>;
 
 /**
  * A structure that provides an association between a reference
@@ -80,7 +80,7 @@ export type QueryGroupFn = (query: Query) => Query;
  *  publisher: 'SportsPublisher'
  * });
  */
-export interface AssociatedReference {
-  ref: CollectionReference;
-  query: Query;
+export interface AssociatedReference<T = DocumentData> {
+  ref: CollectionReference<T>;
+  query: Query<T>;
 }
