@@ -4,13 +4,13 @@ import { map } from 'rxjs/operators';
 
 function _fromRef<T, R>(ref: Reference<T>, scheduler: SchedulerLike = asyncScheduler): Observable<R> {
   return new Observable(subscriber => {
-    let unsubscribe;
+    let unsubscribe: () => void;
     if (scheduler != null) {
       scheduler.schedule(() => {
-        unsubscribe = ref.onSnapshot(subscriber);
+        unsubscribe = ref.onSnapshot({ includeMetadataChanges: true }, subscriber);
       });
     } else {
-      unsubscribe = ref.onSnapshot(subscriber);
+      unsubscribe = ref.onSnapshot({ includeMetadataChanges: true }, subscriber);
     }
 
     return () => {
@@ -21,17 +21,17 @@ function _fromRef<T, R>(ref: Reference<T>, scheduler: SchedulerLike = asyncSched
   });
 }
 
-export function fromRef<R>(ref: DocumentReference | Query, scheduler?: SchedulerLike) {
+export function fromRef<R, T>(ref: DocumentReference<T> | Query<T>, scheduler?: SchedulerLike) {
   return _fromRef<typeof ref, R>(ref, scheduler);
 }
 
-export function fromDocRef<T>(ref: DocumentReference, scheduler?: SchedulerLike): Observable<Action<DocumentSnapshot<T>>> {
-  return fromRef<DocumentSnapshot<T>>(ref, scheduler)
+export function fromDocRef<T>(ref: DocumentReference<T>, scheduler?: SchedulerLike): Observable<Action<DocumentSnapshot<T>>> {
+  return fromRef<DocumentSnapshot<T>, T>(ref, scheduler)
     .pipe(
       map(payload => ({ payload, type: 'value' }))
     );
 }
 
-export function fromCollectionRef<T>(ref: Query, scheduler?: SchedulerLike): Observable<Action<QuerySnapshot<T>>> {
-  return fromRef<QuerySnapshot<T>>(ref, scheduler).pipe(map(payload => ({ payload, type: 'query' })));
+export function fromCollectionRef<T>(ref: Query<T>, scheduler?: SchedulerLike): Observable<Action<QuerySnapshot<T>>> {
+  return fromRef<QuerySnapshot<T>, T>(ref, scheduler).pipe(map(payload => ({ payload, type: 'query' })));
 }
