@@ -1,12 +1,14 @@
 import { BuilderContext, targetFromTargetString } from '@angular-devkit/architect';
-import { BuildTarget, FirebaseTools, FSHost } from '../interfaces';
+import { BuildTarget, DeployBuilderSchema, FirebaseTools, FSHost } from '../interfaces';
 import { existsSync, readFileSync, renameSync, writeFileSync } from 'fs';
 import { copySync, removeSync } from 'fs-extra';
 import { dirname, join } from 'path';
 import { execSync } from 'child_process';
-import { defaultFunction, defaultPackage, NODE_VERSION } from './functions-templates';
+import { defaultFunction, defaultPackage } from './functions-templates';
 import { satisfies } from 'semver';
 import * as open from 'open';
+
+export type DeployBuilderOptions = DeployBuilderSchema | Record<string, string>;
 
 const escapeRegExp = (str: string) => str.replace(/[\-\[\]\/{}()*+?.\\^$|]/g, '\\$&');
 
@@ -19,10 +21,11 @@ const deployToHosting = (
   firebaseTools: FirebaseTools,
   context: BuilderContext,
   workspaceRoot: string,
-  preview: boolean
+  options: DeployBuilderOptions,
+  firebaseToken?: string,
 ) => {
 
-  if (preview) {
+  if (options.preview) {
     const port = 5000; // TODO make this configurable
 
     setTimeout(() => {
@@ -40,7 +43,8 @@ const deployToHosting = (
         return firebaseTools.deploy({
           // tslint:disable-next-line:no-non-null-assertion
           only: 'hosting:' + context.target!.project,
-          cwd: workspaceRoot
+          cwd: workspaceRoot,
+          token: firebaseToken,
         });
       } else {
         return Promise.resolve();
@@ -52,7 +56,8 @@ const deployToHosting = (
     return firebaseTools.deploy({
       // tslint:disable-next-line:no-non-null-assertion
       only: 'hosting:' + context.target!.project,
-      cwd: workspaceRoot
+      cwd: workspaceRoot,
+      token: firebaseToken,
     });
 
   }
@@ -68,10 +73,10 @@ const getVersionRange = (v: number) => `^${v}.0.0`;
 
 const findPackageVersion = (name: string) => {
   const match = execSync(`npm list ${name}`).toString().match(` ${escapeRegExp(name)}@.+\\w`);
-  return match ? match[0].split(`${name}@`)[1] : null;
+  return match ? match[0].split(`${name}@`)[1].split(/\s/)[0] : null;
 };
 
-const getPackageJson = (context: BuilderContext, workspaceRoot: string) => {
+const getPackageJson = (context: BuilderContext, workspaceRoot: string, options: DeployBuilderOptions) => {
   const dependencies = {
     'firebase-admin': 'latest',
     'firebase-functions': 'latest'
@@ -108,7 +113,7 @@ const getPackageJson = (context: BuilderContext, workspaceRoot: string) => {
       });
     }
   } // TODO should we throw?
-  return defaultPackage(dependencies, devDependencies);
+  return defaultPackage(dependencies, devDependencies, options);
 };
 
 export const deployToFunction = async (
@@ -117,14 +122,10 @@ export const deployToFunction = async (
   workspaceRoot: string,
   staticBuildTarget: BuildTarget,
   serverBuildTarget: BuildTarget,
-  preview: boolean,
+  options: DeployBuilderOptions,
+  firebaseToken?: string,
   fsHost: FSHost = defaultFsHost
 ) => {
-  if (!satisfies(process.versions.node, getVersionRange(NODE_VERSION))) {
-    context.logger.warn(
-      `⚠️ Your Node.js version (${process.versions.node}) does not match the Firebase Functions runtime (${NODE_VERSION}).`
-    );
-  }
 
   const staticBuildOptions = await context.getTargetOptions(targetFromTargetString(staticBuildTarget.name));
   if (!staticBuildOptions.outputPath || typeof staticBuildOptions.outputPath !== 'string') {
@@ -152,14 +153,23 @@ export const deployToFunction = async (
   fsHost.moveSync(staticOut, newClientPath);
   fsHost.moveSync(serverOut, newServerPath);
 
+  const packageJson = getPackageJson(context, workspaceRoot, options);
+  const nodeVersion = JSON.parse(packageJson).engines.node;
+
+  if (!satisfies(process.versions.node, getVersionRange(nodeVersion))) {
+    context.logger.warn(
+      `⚠️ Your Node.js version (${process.versions.node}) does not match the Firebase Functions runtime (${nodeVersion}).`
+    );
+  }
+
   fsHost.writeFileSync(
     join(dirname(serverOut), 'package.json'),
-    getPackageJson(context, workspaceRoot)
+    packageJson
   );
 
   fsHost.writeFileSync(
     join(dirname(serverOut), 'index.js'),
-    defaultFunction(serverOut)
+    defaultFunction(serverOut, options)
   );
 
   fsHost.renameSync(
@@ -167,7 +177,7 @@ export const deployToFunction = async (
     join(newClientPath, 'index.original.html')
   );
 
-  if (preview) {
+  if (options.preview) {
     const port = 5000; // TODO make this configurable
 
     setTimeout(() => {
@@ -195,7 +205,8 @@ export const deployToFunction = async (
     return firebaseTools.deploy({
       // tslint:disable-next-line:no-non-null-assertion
       only: `hosting:${context.target!.project},functions:ssr`,
-      cwd: workspaceRoot
+      cwd: workspaceRoot,
+      token: firebaseToken,
     });
   }
 };
@@ -206,9 +217,12 @@ export default async function deploy(
   staticBuildTarget: BuildTarget,
   serverBuildTarget: BuildTarget | undefined,
   firebaseProject: string,
-  preview: boolean
+  options: DeployBuilderOptions,
+  firebaseToken?: string,
 ) {
-  await firebaseTools.login();
+  if (!firebaseToken) {
+    await firebaseTools.login();
+  }
 
   if (!context.target) {
     throw new Error('Cannot execute the build target');
@@ -258,14 +272,16 @@ export default async function deploy(
         context.workspaceRoot,
         staticBuildTarget,
         serverBuildTarget,
-        preview
+        options,
+        firebaseToken,
       );
     } else {
       await deployToHosting(
         firebaseTools,
         context,
         context.workspaceRoot,
-        preview
+        options,
+        firebaseToken,
       );
     }
 
